+++ material/src/material.ts
@@ -1,1 +1,4 @@
 export * from './services/dialogInternalServerErrorRenderer.service';
+export * from './components/dialogInternalServerError/dialogInternalServerError.component';
+export * from './components/dialogInternalServerError/dialogInternalServerError.interface';
+export * from './modules/dialogInternalServerError.module';