+++ material/src/modules/dialogInternalServerError.module.ts
@@ -1,0 +1,17 @@
+import {NgModule} from '@angular/core';
+
+import {DialogInternalServerErrorComponent} from '../components/dialogInternalServerError/dialogInternalServerError.component';
+
+/**
+ * Module used for dialog internal server error component
+ */
+@NgModule(
+{
+    declarations:
+    [
+        DialogInternalServerErrorComponent
+    ]
+})
+export class DialogInternalServerErrorModule
+{
+}