+++ src/internalServerError/internalServerError.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnDestroy, ChangeDetectionStrategy, Inject} from '@angular/core';
+import {Component, OnDestroy, ChangeDetectionStrategy, Inject, ChangeDetectorRef} from '@angular/core';
 import {DomSanitizer} from '@angular/platform-browser';
 import {Subscription} from 'rxjs';
 
@@ -11,7 +11,7 @@ import {InternalServerErrorService, InternalServerErrorInfo, INTERNAL_SERVER_ERR
 {
     selector: "internal-server-error",
     templateUrl: 'internalServerError.component.html',
-    styleUrls: ['internalServerError.component.scss'],
+    styleUrls: ['internalServerError.component.css'],
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class InternalServerErrorComponent implements OnDestroy
@@ -23,23 +23,32 @@ export class InternalServerErrorComponent implements OnDestroy
      */
     private _subscription: Subscription = null;
     
-    //######################### public properties #########################
+    //######################### public properties - template bindings #########################
 
     /**
      * Array of html errors that should be displayed
+     * @internal
      */
     public errorsHtml: InternalServerErrorInfo[] = [];
+
+    /**
+     * Indication whether are errors visible
+     * @internal
+     */
+    public errorsVisible: boolean = false;
 
     //######################### constructor #########################
     constructor(internalServerErrorService: InternalServerErrorService,
                 sanitizerService: DomSanitizer,
-                @Inject(INTERNAL_SERVER_ERROR_RENDERER) private _renderer: InternalServerErrorRenderer)
+                @Inject(INTERNAL_SERVER_ERROR_RENDERER) private _renderer: InternalServerErrorRenderer,
+                private _changeDetector: ChangeDetectorRef)
     {
         this._subscription = internalServerErrorService.internalServerErrorOccured.subscribe((itm: InternalServerErrorInfo) =>
         {
             itm.errorHtml = sanitizerService.bypassSecurityTrustResourceUrl('data:text/html;charset=utf-8,' + encodeURI(<string>itm.errorHtml).replace(/#/g, "%23"));
 
             this.errorsHtml.push(itm);
+            _changeDetector.detectChanges();
         });
     }
 
@@ -77,5 +86,12 @@ export class InternalServerErrorComponent implements OnDestroy
         {
             this.errorsHtml.splice(index, 1);
         }
+
+        if(!this.errorsHtml.length)
+        {
+            this.errorsVisible = false;
+        }
+
+        this._changeDetector.detectChanges();
     }
 }